added Long/Int support for resource amount

+++ core/src/main/kotlin/net/asch/bulkit/common/data/extensions/FluidStackExtension.kt
@@ -5,4 +5,4 @@ import net.minecraft.world.level.material.Fluid
 import net.neoforged.neoforge.fluids.FluidStack
 
 fun FluidStack.identifier(): ResourceIdentifier<Fluid> = ResourceIdentifier(fluidHolder, componentsPatch)
-fun ResourceIdentifier<Fluid>.of(amount: Long): FluidStack = FluidStack(resource, amount.toInt(), dataComponents)
+fun ResourceIdentifier<Fluid>.of(amount: Int): FluidStack = FluidStack(resource, amount, dataComponents)
